* bug fix in DriftDiffusionProperties

+++ src/core/modules/drift-diffusion/DriftDiffusionProperties.C
@@ -376,11 +376,15 @@ DriftDiffusionProperties::do_init_alloy(const PhysicalModelInterface* comp_A,
       scB->_pyropolarization, xa);
 
 
-  PhysicalModelInterface::destroy(_thermoelectric_power);
-  _thermoelectric_power = create_submodel_copy(scA->_thermoelectric_power);
-  _thermoelectric_power->set_driftdiffusionproperties(this);
-  _thermoelectric_power->init_alloy(scA->_thermoelectric_power,
-      scB->_thermoelectric_power, xa);
+  if (scA->_thermoelectric_power != NULL)
+  {
+    assert(scB->_thermoelectric_power != NULL);
+    PhysicalModelInterface::destroy(_thermoelectric_power);
+    _thermoelectric_power = create_submodel_copy(scA->_thermoelectric_power);
+    _thermoelectric_power->set_driftdiffusionproperties(this);
+    _thermoelectric_power->init_alloy(scA->_thermoelectric_power,
+        scB->_thermoelectric_power, xa);
+  }
 
 }
